Правка задания Калькулятор ООД

+++ chapter_007/src/main/java/ru/job4j/ood/warehouse/template/Beef.java
@@ -1,10 +1,9 @@
 package ru.job4j.ood.warehouse.template;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Beef extends Food {
-    public Beef(String title, LocalDate expireDate, LocalDate createDate, BigDecimal price, BigDecimal discount) {
+    public Beef(String title, LocalDate expireDate, LocalDate createDate, int price, int discount) {
         super(title, expireDate, createDate, price, discount);
     }
